Prevent short circuiting the shouldSort() function
This removes a short circuit that would previously prevent checking
shouldSort() when the execution path begins from the PLAYER_REGEN_ENABLED
event.

By removing this short circuit we can successfully re-sort frames in the
event of any changes outside the context of GROUP_ROSTER_UPDATE event
and updates to the Cell layout.